requirements.txt
# Core dependencies
faiss-cpu==1.7.4
numpy>=1.26.0
httpx>=0.27.0
PyYAML>=6.0.1

# HTML/Markdown parsing & conversion
beautifulsoup4>=4.12.2
lxml>=5.2.0
markdown-it-py>=3.0.0

# Logging / CLI improvements (optional but useful)
rich>=13.7.0

# For scraping (if you run the ocr/html scrapers)
requests>=2.31.0

# OCR support (optional; if you ever use OCR again)
pytesseract>=0.3.10
Pillow>=10.0.0

# Development helpers
black>=24.3.0        # code formatter
isort>=5.12.0        # import sorter